FossilOrigin-Name: 25a6c00b828ac6bfd45fdbc1bb85b1ca2565bfcbf0bb580e3905446aed811975

+++ src/OFASN1BitString.m
@@ -172,9 +172,9 @@
 	return true;
 }
 
-- (uint32_t)hash
+- (unsigned long)hash
 {
-	return _bitStringValue.hash + (uint32_t)_bitStringLength;
+	return _bitStringValue.hash + (unsigned long)_bitStringLength;
 }
 
 - (OFString *)description
