+++ src/OFASN1IA5String.m
@@ -112,7 +112,7 @@
 	return true;
 }
 
-- (uint32_t)hash
+- (unsigned long)hash
 {
 	return _IA5StringValue.hash;
 }
