+++ src/OFASN1Integer.m
@@ -113,9 +113,9 @@ of_asn1_der_integer_parse(const unsigned char *buffer, size_t length)
 	return true;
 }
 
-- (uint32_t)hash
+- (unsigned long)hash
 {
-	return (uint32_t)_longLongValue;
+	return (unsigned long)_longLongValue;
 }
 
 - (OFString *)description
